+++ src/main/java/com/capstone/dto/request/RecalculateProgressRequestDto.java
@@ -1,4 +1,4 @@
-package com.capstone.dto.roadmap;
+package com.capstone.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,10 +11,7 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AtomProgressRequestDto {
+public class RecalculateProgressRequestDto {
     UUID learnerId;
-    UUID atomId;
-    UUID capsuleId;
-    UUID trackId;
     UUID talentRouteId;
 }
