+++ src/main/java/com/capstone/service/LearnerProgressService.java
@@ -1,7 +1,7 @@
 package com.capstone.service;
 
-import com.capstone.dto.roadmap.AtomProgressRequestDto;
-import com.capstone.dto.roadmap.RecalculateProgressRequestDto;
+import com.capstone.dto.request.AtomProgressRequestDto;
+import com.capstone.dto.request.RecalculateProgressRequestDto;
 
 public interface LearnerProgressService {
     void startAtomProgress(AtomProgressRequestDto atomProgressRequestDto);
