+++ src/main/java/com/capstone/service/LearnerRoadmapService.java
@@ -4,7 +4,7 @@ import com.capstone.dto.response.LearnerAtomProgressDto;
 import com.capstone.dto.response.LearnerCapsuleProgressDto;
 import com.capstone.dto.response.LearnerRoadmapWithProgressDto;
 import com.capstone.dto.response.LearnerTrackProgressDto;
-import com.capstone.dto.roadmap.RoadmapRequestDto;
+import com.capstone.dto.request.RoadmapRequestDto;
 
 import java.util.List;
 import java.util.Optional;
